Update: Modularise Statistic component
