+++ src/components/alexandria/Statistic/component.spec.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-
-import StatisticComponent from '../../../src/components/alexandria/StatisticComponent';
+import StatisticComponent from './component';
 
 describe('StatisticComponent', () => {
   it('should render with value and label', () => {
