+++ src/components/alexandria/Statistic/styles.scss
@@ -1,6 +1,6 @@
-@import '../color';
-@import '../font-size';
-@import '../font-weight';
+@import 'styles/color';
+@import 'styles/font-size';
+@import 'styles/font-weight';
 
 .statistic-component {
   font-size: $font-size-subheader;
